Refactoring packages structure
